feat: keywordSaveToLocal function to utils
- make a util component
