+++ src/components/Search/index.tsx
@@ -28,6 +28,7 @@ import {
 import { useAppDispatch, useAppSelector } from '@/store/hooks';
 import { modalActions } from '@/store/modal/modal-slice';
 import { keywordSuggestActions } from '@/store/searchkeyword/keyword-slice';
+import { KeywordSaveToLocalStorage } from '@/utils';
 
 const SearchBox = () => {
   const dispatch = useAppDispatch();
@@ -51,16 +52,6 @@ const SearchBox = () => {
 
   const KeywordSaveToRedux = (str: string) =>
     dispatch(keywordSuggestActions.handleSearchKeyword(str));
-
-  const KeywordSaveToLocalStorage = (keyword: string) => {
-    let recentSearch: any = localStorage.getItem('recentSearch');
-    recentSearch = recentSearch === null ? [] : JSON.parse(recentSearch);
-    if (recentSearch.length > 5) {
-      recentSearch = recentSearch.slice(0, 5);
-    }
-    recentSearch = new Set([keyword, ...recentSearch]);
-    localStorage.setItem('recentSearch', JSON.stringify([...recentSearch]));
-  };
 
   const onSubmit = (e: any) => {
     e.preventDefault();
