+++ src/components/index.ts
@@ -3,7 +3,6 @@ export { default as Restaurants } from './Restaurants';
 export { default as CategoryItem } from './CategoryItem';
 export { default as Category } from './Category';
 export { default as Carousel } from './Carousel';
-export { default as SearchBox } from './Search';
 export { default as Banner } from './Banner';
 export { default as Header } from './Header';
 export { default as Modal } from './Modal';
@@ -14,7 +13,7 @@ export { default as BestRestaurantItem } from './BestRestaurantItem';
 export { default as MoreButton } from './MoreButton';
 export { default as LinkCopyButton } from './LinkCopyButton';
 export { default as SurroundPopluars } from './SurroundPopulars';
-export { default as SearchModal } from './Search';
+export { default as SearchBox } from './Search';
 export { default as Portal } from './Portal';
 export { default as Reviews } from './Reviews';
 export { default as ReviewWrite } from './ReviewWrite';
