+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { getErrorMessage } from './getErrorMessage';
+export { KeywordSaveToLocalStorage } from './keywordSaveToLocal';
